+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.99.1
 
+- Swift, Node: Don't log an error about unexpected server timestamps on unauthenticated connections.